+++ docs/conf.py
@@ -7,7 +7,7 @@
 project = "uiprotect"
 copyright = "2024, UI Protect Maintainers"
 author = "UI Protect Maintainers"
-release = "3.6.0"
+release = "4.2.0"
 
 # General configuration
 extensions = [