+++ src/uiprotect/api.py
@@ -2001,7 +2001,6 @@ class ProtectApiClient(BaseApiClient):
         except (
             TimeoutError,
             asyncio.TimeoutError,
-            asyncio.CancelledError,
             aiohttp.ServerDisconnectedError,
             client_exceptions.ClientError,
         ) as err: